+++ mxops/utils/chain_simulator.py
@@ -16,10 +16,10 @@ from multiversx_sdk.network_providers.errors import GenericError
 from mxops.config.config import Config
 from mxops.data import data_path
 from mxops.utils.logger import get_logger
-from mxops.enums import NetworkEnum
+from mxops.enums import LogGroupEnum, NetworkEnum
 from mxops.common.providers import MyProxyNetworkProvider
 
-LOGGER = get_logger("CHAIN SIMULATOR")
+LOGGER = get_logger(LogGroupEnum.MSC)
 
 
 def add_subparser(subparsers_action: _SubParsersAction):