+++ docs/source/dev_documentation/changelog.md
@@ -35,6 +35,7 @@
 - SceneStep can now be repeated
 - SceneStep can now target a folder of scenes
 - User can now evaluate an expression with the data CLI
+- logs are now saved to the MxOps data folder and scenario folders
 
 ### Changed